+++ R/powrballad-package.r
@@ -3,3 +3,10 @@
 #' @name powrballad
 #' @docType package
 NULL
+
+
+
+
+#' A list of Power Ballads that can be used to randomly select from
+#'
+"ballads"