+++ NEWS.md
@@ -1,6 +1,8 @@
 # powrballad 0.0.0.9000
 
 * Added a `NEWS.md` file to track changes to the package.
-
-
+* Added a select_powerballad() function that randomly selects a song
+* Added a test
+* Added contributors
+* Added News File